refactor: 1차 피드백 내용 적용

+++ src/main/java/lotto/dto/response/ResultResponse.java
@@ -1,14 +1,15 @@
 package lotto.dto.response;
 
+import java.math.BigDecimal;
 import lotto.domain.Rank;
 
 import java.util.Map;
 
 public record ResultResponse(
-    Map<Rank, Integer> rankCount,
-    double rateOfReturn
+        Map<Rank, Long> rankCount,
+        BigDecimal rateOfReturn
 ) {
-    public static ResultResponse of(Map<Rank, Integer> rankCount, double rateOfReturn) {
+    public static ResultResponse of(Map<Rank, Long> rankCount, BigDecimal rateOfReturn) {
         return new ResultResponse(rankCount, rateOfReturn);
     }
 }
